Move instructions to submodule
