+++ emulator/src/core/emulator/instructions/mod.rs
@@ -1,0 +1,1 @@
+mod arithmetic;
